test(app): raise coverage, trim lcov excludes, split widgetbook (#1177)
- Move Widgetbook catalog to lib/widgetbook/catalog.dart; thin widgetbook.dart entry
- Fix availableWorkTargetsProvider deduplication in games_provider
- HiveBoxNames as abstract final class for coverage-friendly constants
- Add tests: GameService cache, map/workflow providers, split fleet, widgetbook smoke
- Extend province overlay, dialogue, game screen overture tests
- check_coverage_threshold.sh: shorter exclude list; regenerate filtered lcov

Made-with: Cursor
