+++ app/lib/config/constants.dart
@@ -1,7 +1,5 @@
 /// App and Hive constants. TDD 15 Local Storage box names.
-class HiveBoxNames {
-  HiveBoxNames._();
-
+abstract final class HiveBoxNames {
   static const String settings = 'settings';
   static const String games = 'games';
   static const String offlineQueue = 'offline_queue';
