+++ app/test/game_screen_overture_pending_test.dart
@@ -61,4 +61,3 @@ void main() {
     expect(find.text('Submit'), findsOneWidget);
   });
 }
-
