+++ app/test/game_service_cache_test.dart
@@ -2,6 +2,9 @@ import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'dart:io';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_logic/colonizethis_logic.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_save/colonizethis_save.dart';
 import 'package:hive/hive.dart';
 
@@ -32,6 +35,64 @@ void main() {
     test('getMapData returns null for unknown game id', () {
       final result = service.getMapData('no-such-game');
       expect(result, isNull);
+    });
+
+    test('fresh GameService loads map from storage on getMapData and loadGame', () {
+      const gameId = 'persist_map';
+      final config = GameSetupConfig(
+        selectedGreatPowerIds: ['england'],
+        continentCount: 1,
+        minorNationCount: 0,
+        tribeCount: 1,
+        numProvincesOldWorld: 3,
+        numProvincesNewWorld: 2,
+      );
+      final writer = GameService(box, GameSaveAdapter());
+      writer.createNewGame(id: gameId, config: config);
+
+      final reader = GameService(box, GameSaveAdapter());
+      final first = reader.getMapData(gameId);
+      expect(first, isNotNull);
+      expect(first!.combinedTopology, isNotNull);
+
+      final second = reader.getMapData(gameId);
+      expect(second, isNotNull);
+
+      final freshForLoad = GameService(box, GameSaveAdapter());
+      final loaded = freshForLoad.loadGame(gameId);
+      expect(loaded, isNotNull);
+    });
+
+    test('runTurnResolution uses merge when aiOrders is non-null', () {
+      const gameId = 'merge_ai';
+      final config = GameSetupConfig(
+        selectedGreatPowerIds: ['england'],
+        continentCount: 1,
+        minorNationCount: 0,
+        tribeCount: 1,
+        numProvincesOldWorld: 3,
+        numProvincesNewWorld: 2,
+      );
+      final g = service.createNewGame(id: gameId, config: config);
+      final result = service.runTurnResolution(
+        g,
+        orders: const Orders(),
+        aiOrders: const Orders(),
+      );
+      expect(result, isA<TurnResolutionComplete>());
+    });
+
+    test('createNewGame assigns generated id when id omitted', () {
+      final config = GameSetupConfig(
+        selectedGreatPowerIds: ['england'],
+        continentCount: 1,
+        minorNationCount: 0,
+        tribeCount: 1,
+        numProvincesOldWorld: 3,
+        numProvincesNewWorld: 2,
+      );
+      final g = service.createNewGame(config: config);
+      expect(g.id, startsWith('game_'));
     });
   });
 }
