+++ app/test/themes_and_widgetbook_test.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app/config/themes.dart';
-import 'package:colonizethis_app/widgetbook.dart' as ct_widgetbook;
 
 void main() {
   suppressLogsForTests();
